Align classroom solutions with the concepts taught in each topic
Replace premature utility APIs with taught loops, arithmetic and character operations. Document prerequisites and input domains for all 37 exercises, align model contracts with the handouts, and clarify that the Topic 1 benchmark is an instructor demonstration.

Keep every exercise self-contained in the default package. Verify all examples compile independently and run 67 classroom input/output cases plus API boundary checks.

+++ Topic-3/08_Numeric_Palindrome/NumericPalindrome.java
@@ -18,7 +18,7 @@ public class NumericPalindrome {
 
     public static boolean isPalindrome(int number) {
         int original = number;
-        int reversed = 0;
+        long reversed = 0;
 
         while (number > 0) {
             int digit = number % 10;
